// client/src/router/enums.ts
export enum RouteNames {
  Home = 'Home',
  Details = 'Details',
  Holidays = 'Holidays',
  Sickness = 'Sickness',
  Payslips = 'Payslips',
  Login = 'Login',
}

export enum RoutePaths {
  Home = '/',
  Details = '/details',
  Holidays = '/holidays',
  Sickness = '/sickness',
  Payslips = '/payslips',
  Login = '/login',
}